app/requirements.txt
fastapi
uvicorn
redis
aio-pika
python-dotenv
pydantic